+++ starlight-basic/src/main/java/org/atcraftmc/starlight/display/CustomJoinMessage.java
@@ -1,4 +1,0 @@
-package org.atcraftmc.starlight.display;
-
-public class CustomJoinMessage {
-}