+++ changes.md
@@ -1,3 +1,17 @@
+## 26.2.1
+- BuildTime and Version are now correct.
+- Fixed chat error on reloading.
+- Add module `sit-on-player`.
+- Add module `player-join-message`
+- Add `en_us` translation for `starlight-display`
+- Add `en_us` translation for `starlight-core`
+- Add `en_us` translation for `starlight-warps`
+- Add `en_us` translation for `starlight-tweaks`
+- Add `en_us` translation for `starlight-oddities`
+- Fixed crashing by duplicate reloading.
+- Fixed custom block rendering on pickup.
+
+
 ## 26.2.0
 - Fixed library loading exception in some paths.
 - Elevator can now recognize legacy blocks.